requirements.txt
mysql-connector-python
passlib
requests
minecraft-launcher-lib